+++ lib/warpdrived/register.go
@@ -1,4 +1,4 @@
-package server
+package warpdrived
 
 import (
 	"fmt"
