+++ lib/warpdrived/service/offer.go
@@ -3,11 +3,11 @@ package service
 import (
 	"context"
 	"encoding/json"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/core"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/notify"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage/file"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage/memory"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/core"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/notify"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage/file"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage/memory"
 	"github.com/cryptopunkscc/astrald/proto/warpdrive"
 	"sort"
 	"sync"
