+++ lib/warpdrived/service/peer.go
@@ -1,9 +1,9 @@
 package service
 
 import (
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/core"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage/file"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage/memory"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/core"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage/file"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage/memory"
 	"github.com/cryptopunkscc/astrald/proto/contacts"
 	"github.com/cryptopunkscc/astrald/proto/warpdrive"
 )
