+++ lib/warpdrived/storage/file/offers.go
@@ -2,8 +2,8 @@ package file
 
 import (
 	"encoding/gob"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/core"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/core"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage"
 	"github.com/cryptopunkscc/astrald/proto/warpdrive"
 	"io/fs"
 	"log"
