+++ lib/warpdrived/storage/file/storage.go
@@ -1,8 +1,8 @@
 package file
 
 import (
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/core"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/core"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage"
 	"io"
 	"os"
 	"path/filepath"
