+++ lib/warpdrived/storage/memory/offers.go
@@ -1,8 +1,8 @@
 package memory
 
 import (
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/core"
-	"github.com/cryptopunkscc/astrald/cmd/warpdrived/storage"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/core"
+	"github.com/cryptopunkscc/astrald/lib/warpdrived/storage"
 	"github.com/cryptopunkscc/astrald/proto/warpdrive"
 )
